fix(seo): scope ProfilePage JSON-LD to home page only
