+++ src/app/page.tsx
@@ -1,5 +1,11 @@
 import Home from '@/components/home/home';
+import { HomeJsonLd } from '@/components/seo/home-jsonld';
 
 export default function HomePage() {
-  return <Home />;
+  return (
+    <>
+      <HomeJsonLd />
+      <Home />
+    </>
+  );
 }
